requirements.txt
netaddr
pynetbox
python-nmap
